temp

+++ Assets/SdpTransform/Models/Attributes/RtcpFb.cs
@@ -7,7 +7,7 @@ namespace Utilme.SdpTransform
     public class RtcpFb
     {
         public const string Label = "rtcp-fb:";
-        public int PayloadType { get; set; }
+        public byte PayloadType { get; set; }
         public string Type { get; set; }
         public string SubType { get; set; }
