+++ Assets/SdpTransform/Models/Attributes/Rtpmap.cs
@@ -7,9 +7,9 @@ namespace Utilme.SdpTransform
     public class Rtpmap
     {
         public const string Label = "rtpmap:";
-        public int PayloadType { get; set; }
+        public byte PayloadType { get; set; }
         public string EncodingName { get; set; }
-        public int ClockRate { get; set; }
-        public int? Channels { get; set; }
+        public uint ClockRate { get; set; }
+        public byte? Channels { get; set; }
     }
 }
